moved `useSpotlightActions` logic back in `ChatModelSpotlight`
